% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server_func}
\alias{server_func}
\title{Serverdelen af dashboardet}
\usage{
server_func(input, output)
}
\arguments{
\item{input}{internal}

\item{output}{internal}
}
\value{
Server-funktion til Dashboardet
}
\description{
Serverdelen af dashboardet
}
